requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asn1crypto @ file:///C:/ci/asn1crypto_1652344143274/work
async-generator==1.10
attrs==22.1.0
backcall==0.2.0
bcrypt==4.0.1
bleach==4.1.0
brotlipy==0.7.0
certifi @ file:///C:/b/abs_ac29jvt43w/croot/certifi_1665076682579/work/certifi
cffi @ file:///C:/b/abs_49n3v2hyhr/croot/cffi_1670423218144/work
charset-normalizer==2.0.12
cloudpickle @ file:///tmp/build/80754af9/cloudpickle_1632508026186/work
colorama==0.4.5
cryptography==36.0.2
cycler==0.11.0
dataclasses==0.8
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
filelock==3.8.2
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
greenlet==2.0.1
idna @ file:///C:/b/abs_bdhbebrioa/croot/idna_1666125572046/work
importlib-metadata==4.8.3
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.7.2
jedi==0.17.2
Jinja2==3.0.3
joblib==1.1.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186184278/work
mkl-service==2.4.0
mpmath==1.2.1
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.6
nose==1.3.7
notebook==6.4.10
numpy==1.23.5
oscrypto==1.3.0
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
paramiko==2.12.0
parso==0.7.1
patsy==0.5.3
pickleshare==0.7.5
Pillow==8.4.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==3.19.6
pyarrow==8.0.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodomex==3.16.0
Pygments==2.13.0
PyJWT @ file:///C:/ci/pyjwt_1657529430378/work
PyNaCl==1.5.0
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
pyrsistent==0.18.0
PySocks @ file:///C:/ci/pysocks_1605287845585/work
python-dateutil==2.8.2
pytz==2022.6
pywin32==305
pywinpty==1.1.6
pyzmq==24.0.1
qtconsole==5.2.2
QtPy==2.0.1
requests @ file:///C:/ci/requests_1657717096906/work
rexmex==0.1.2
scikit-learn==1.0.2
scipy==1.4.1
seaborn==0.11.2
Send2Trash==1.8.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
snowflake-connector-python==2.8.3
snowflake-snowpark-python @ file:///C:/b/abs_c9l6p0q60d/croot/snowflake-snowpark-python_1667400726103/work
snowflake-sqlalchemy==1.4.3
SQLAlchemy==1.4.44
statsmodels==0.12.2
sympy==1.11.1
terminado==0.12.1
testpath==0.6.0
threadpoolctl==3.1.0
tornado==6.1
traitlets==4.3.3
typing_extensions @ file:///C:/b/abs_89eui86zuq/croot/typing_extensions_1669923792806/work
urllib3 @ file:///C:/b/abs_21qa0j12xt/croot/urllib3_1670527000672/work
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.1
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306167264/work
wincertstore==0.2
zipp==3.6.0
